backend/requirements.txt
aiofiles==23.1.0
fastapi==0.100.0
uvicorn[standard]==0.23.2
aiofiles==23.1.0
pytest==7.4.0
requests==2.31.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.0
